[sec1] Solucionado el tema del short

+++ sec_1/Clase_220818/src/main/App.java
@@ -4,13 +4,15 @@ import model.Cancion;
 
 public class App {
 
+    // https://stackoverflow.com/questions/477750/primitive-type-short-casting-in-java
+    // https://docs.oracle.com/javase/specs/jls/se7/html/jls-4.html#jls-4.2.2
     public static void main(String[] args) {
         Cancion can = new Cancion(
             "Bandolero",
             "Don Omar",
             "Los Bandoleros",
-            2005,
-            320,
+            (short)2005,
+            (short)320,
             "Reggaeton",
             "Tego Calderón"
         );
